fix: manually renamed the package in the model files
