+++ app/src/main/java/at/aau/monopoly/klagenfurt/model/field/GoField.kt
@@ -1,6 +1,6 @@
-package at.aau.monopoly.websocketdemoserver.model.field
+package at.aau.monopoly.klagenfurt.model.field
 
-import at.aau.monopoly.websocketdemoserver.model.enums.FieldType
+import at.aau.monopoly.klagenfurt.model.enums.FieldType
 
 data class GoField(
     override val id: Int = 0,
